# backend/docs/swagger.yaml
basePath: /api
definitions:
  models.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      request_id:
        type: string
    type: object
...
  models.SanctumRequest:
    properties:
      created_at:
        type: string
      id:
        type: integer
      reason:
        type: string
      requested_by_user:
        $ref: '#/definitions/models.User'
      requested_by_user_id:
        type: integer
      requested_name:
        type: string
      requested_slug:
        type: string
      review_notes:
        type: string
      reviewed_by_user:
        $ref: '#/definitions/models.User'
      reviewed_by_user_id:
        type: integer
      status:
        $ref: '#/definitions/models.SanctumRequestStatus'
      updated_at:
        type: string
    type: object
...
  models.StreamMessage:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      stream_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  server.CreateStreamRequest:
    properties:
      category:
        type: string
      description:
        type: string
      stream_type:
        enum:
        - youtube
        - twitch
        - hls
        - iframe
        type: string
      stream_url:
        type: string
      tags:
        type: string
      thumbnail_url:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - stream_type
    - stream_url
    - title
    type: object
  server.SanctumDTO:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      default_chat_room_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/models.SanctumStatus'
      updated_at:
        type: string
    type: object
  server.SanctumMembershipDTO:
    properties:
      created_at:
        type: string
      role:
        $ref: '#/definitions/models.SanctumMembershipRole'
      sanctum:
        $ref: '#/definitions/server.SanctumDTO'
      sanctum_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  server.UpdateStreamRequest:
    properties:
      category:
        type: string
      description:
        type: string
      stream_type:
        type: string
      stream_url:
        type: string
      tags:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
    type: object
host: localhost:8375
info:
  contact:
    email: support@sanctum.dev
    name: API Support
  description: Social media platform API with posts, comments, messaging, and friends
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sanctum API
  version: "1.0"
paths:
  /admin/sanctum-requests:
    get:
      description: List sanctum requests by status. Defaults to pending.
      parameters:
      - description: Filter status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SanctumRequest'
            type: array
      security:
      - BearerAuth: []
      summary: List sanctum requests for admins
      tags:
      - sanctums-admin
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              refresh_token:
                type: string
              token:
                type: string
              user:
                $ref: '#/definitions/models.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token and blacklist current access token
      parameters:
      - description: Logout request
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      summary: User logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              refresh_token:
                type: string
              token:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Refresh JWT token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Signup request
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              refresh_token:
                type: string
              token:
                type: string
              user:
                $ref: '#/definitions/models.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      summary: User signup
      tags:
      - auth
...
  /sanctums/{slug}:
    get:
      description: Fetch active sanctum detail by slug.
      parameters:
      - description: Sanctum slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SanctumDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get sanctum by slug
      tags:
      - sanctums
  /sanctums/memberships/bulk:
    post:
      consumes:
      - application/json
      description: Upsert current user's followed sanctums by slug and keep owner/mod
        memberships intact.
      parameters:
      - description: Sanctum follow list
        in: body
        name: request
        required: true
        schema:
          properties:
            sanctum_slugs:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.SanctumMembershipDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Save sanctums I follow
      tags:
      - sanctums
  /sanctums/memberships/me:
    get:
      description: List sanctum memberships for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.SanctumMembershipDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Get my sanctum memberships
      tags:
      - sanctums
  /sanctums/requests:
    post:
      consumes:
      - application/json
      description: Submit a request for a new sanctum.
      parameters:
      - description: Sanctum request
        in: body
        name: request
        required: true
        schema:
          properties:
            reason:
              type: string
            requested_name:
              type: string
            requested_slug:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SanctumRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create sanctum request
      tags:
      - sanctums
  /sanctums/requests/me:
    get:
      description: List sanctum requests submitted by the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SanctumRequest'
            type: array
      security:
      - BearerAuth: []
      summary: Get my sanctum requests
      tags:
      - sanctums
  /streams:
    get:
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get live streams
      tags:
      - Streams
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream data
        in: body
        name: stream
        required: true
        schema:
          $ref: '#/definitions/server.CreateStreamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Stream'
      summary: Create a new stream
      tags:
      - Streams
  /streams/{id}:
    delete:
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a stream
      tags:
      - Streams
    get:
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stream'
      summary: Get stream by ID
      tags:
      - Streams
    put:
      consumes:
      - application/json
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stream data
        in: body
        name: stream
        required: true
        schema:
          $ref: '#/definitions/server.UpdateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stream'
      summary: Update a stream
      tags:
      - Streams
  /streams/{id}/end:
    post:
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stream'
      summary: End streaming
      tags:
      - Streams
...
  /streams/{id}/messages:
    get:
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StreamMessage'
            type: array
      summary: Get stream chat messages
      tags:
      - Streams
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StreamMessage'
      summary: Send stream chat message
      tags:
      - Streams
...
  /ws/ticket:
    post:
      consumes:
      - application/json
      description: Generates a short-lived (60s), single-use ticket for WebSocket
        authentication to avoid sending JWTs in query parameters.
      produces:
      - application/json
      responses:
        "200":
          description: 'ticket: string, expires_in: int'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 'error: Redis not available'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Issue WebSocket ticket
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"